Feature/dispatcher (#2)
* Feat: Added a function to make a bitmask for a parameter pack

* Feat: Added overload to Entity::has_component to take a bitmask

Used the overload in the test code

* Feat: Finished the exec function for systems

Running a system on all components with all the correct types can now be done with one function call to execute().

* Feat: Separated World classes into separate hpp files. Started documentation

* Feat: Added WorldBuilder docs

* Feat: Added docs and cleaned up world_class

* Feat: Added docs and cleaned up Entity class

* Feat: Added Doxyfile to use doxygen

* Feat: Added Executable class

The idea is that the dispatcher will be a container of Executable objects, for which exec() can then be called in order (and ideally in parallel)

* Feat: Added a vector of Executables which can be iterated over

This is the first major step to having a dispatcher, because now the template types for each system can be hidden and aren't required at runtime, because of how System is able to implement Executable::exec(). This will now make it possible to add anything which inherits System to the dispatcher, and then later add ways to describe dependencies in order to determine what can be executed in parallel.

* Feat: Added api for adding systems and dispatching the systems

This will probably do for the dispatcher for now, I'll add brief documentation then merge the branch. A lot of work has been done, and now some more thourough testing should be done since I have the basics working

* Feat: Initial dispatcher docs

* Feat: Added documentation to the RegistryNode.

I mainly focused on the safety of using the void* and how I ensure that the data is only ever cast to the proper type through the use of the typeid hash_code. So far this seems to work properly, however it's probably worth sitting down with the prof to make sure that the invariants are upheld

* Feat: Added docs for the System class

* Feat: Added const to some of the example systems to prove it works