Fix: Const arguments seemed to break things today
Not sure what changed, but after merging, and cloning to the VM, the code no longer compiled, despite the fact that I had tested the code before committing it. The error themselves are a little wierd in that they're saying im making a vector<const Position> for instance, when it really should be a vector<tuple<const* Position>>, which I dont think is an issue. Anyways for the moment, i've removed the const qualifiers, but I'll return to this in the future