Feature/system batching (#4)
* Feat: Added a dependency builder

I used a BFS topological sort to determine which systems can be run in parallel. The algorithm is as follows: 1. Create a graph from the input parameters. Adding a system now requires that any dependencies are indicated. 2. Add all systems with incoming dependencies to a 'stage'. 3. Remove the noted systems from the graph. 4. Repeat stages 2&3 until the graph is empty. If the graph is not empty, and no systems have been added to a stage, then the graph has at least one cycle, and an order cannot be determined. This throws a runtime error.

* Feat: Improved add_system api, and fixed old example

* Feat: Added multithreading to dispatcher

* Feat: Added some comments

* Feat: Added documentation

* Feat: Ensured that every Entity is linked to it's own Entity Component