Feat: Added a NodeType to the registryNode.
This will ultimately allow for slightly different behaviour for Components and Resources all wrapped behind the RegistryNode type. This also allows for another safety check to ensure that the RegistryNode constructor wasn't called, by assigning the NodeType to Unkonwn in the constructor, and is set with the create(NodeType) function.