Feat: Systems can now use resources
Systems can include a Resource type in the system definition, which exposes the resources to the System's run function. This allows for the creation of entities at runtime. In order for this to be threadsafe, the User must ensure that any system which uses the World* resource executes alone. If this is not the case then there is a possibility of race conditions.