Feat: Solved use after free error, and fixed problem with removing components from entities
I believe the use after free was a result of an automatic resize in the underlying vector of the RegistryNode. The fetch<Ts ...> retured a set of pointers to the inside of the vector, but as more entities were added, the vector resize invalidated the pointers created. This is why calling e->eid() was causing problems but separating the creation from accessing causes no problems. Additionally, there was a problem with removing components. The component would be erased, but the Entity would never actually delete it's knowledge of the component. This was fixed by having the remove lambda call the entities remove_component.