Pong (#8)
* Feat: Started making pong for demonstration

* Feat: Added user input to control paddles

* Feat: Added collision between balls and paddles

* Feat: Added score

* Feat: Added a ball spawning system & changed Paddle to Side

* Fix: Separated pong files into different headers

* Fix: Separated resources from components

* Fix: wouldn't compile b/c forgot to include pong/resources to pong.cpp

* Feat: Added some commments explaining the Pong Example"

* Feat: Generalized the Text drawing system

* Feat: Better collision. Fix: Seg fault on entity removal

There was a bug with removing Entities, where removing the Entity Componenet influenced the pointers used in other remove lambda functions. This was solved by sorting the lambda functions by decreasing cid, then by decreasing idx. This is ok, because ALL removed components of the same type will still be removed in index decreasing order, but it ensures that Entities are the last Component removed, thus removing the issue.